pyproject.toml
[project]
name = "csrankingsarena"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.11",
    "marker-pdf>=1.6.2",
    "pandas>=2.2.3",
    "ruff>=0.11.5",
]
